Complete the requested changes and add two new sections projects and gallery to the index.html
I update the index.html and change the tags of the intro paragraph from <h2> to <p>. Also, I update the id of my personal picture and
improve the style of my personal picture with the figure tag. In this commit, I also add two more sections to the index.html for the
features to be added in the next PR. style.css file is updated to adapt the changes I've made in index.html.