Address the request changes
Update the gallery and index html file to add alt text for images that describe picture for people using screen readers.
Also, I change the context of the link to my personal gallery to be informative to people using screen readers and remove
unnecessary a tags. Finally, I replace the < and > signs to &lt and &gt.
The gallery_style.css file is updated to adjust the color choice of the prev and next button.