Complete the second step of the Java Servlets Tutorial
Update the index.html file to add a button and a text box such that when the button is clicked, it will call
getMessageFromFetch() function that I added to script.js. The function uses the fetch() function to get response
from the DataServlet.java file and add it to the textbox and display.