Complete the fourth step of the week-3 tutorial and add the feature of comment box to my personal website
Update the DataServlet.java file to add a doPost() function, which makes it possible to take input from the POST
request. Add a form attribute to the index.html to allow the visitors to add comments and by submitting it, the
doPost function will be called and add the comment message to the private list. In the doPost() function, I add
a feature that the visitor will be redirected to the index.html, and onloading the website, the getCommentFromFetch()
function will be called, which will fetch response from the doGet() function and add it to the comment container.
In this way, I am able to show all the comments added.