Address the requested changes in the index.html and script.js
Update the script.js file to use the for each loop while looping through all the comments recorded.
Update the index.html file to use a placeholder attribute to replace the original text for the textarea
attribute in the form. In this way, the user do not need to delete the "Your Comments" text when they want
to input comments.