Add a new feature that all comments are stored permanently in Datastore
Update the doPost() function in DataServlet.java file to store the newly input comment to Datastore. Update the doGet() function
to fetch stored comments from Datastore each time the website page is loaded. The script.js file is updated to adapt the changes
made in DataServlet.java file such that the reponses now are an objects that contain comments instead of strings. Comment.java file
is created to provide a data structure that I can store the comment content and its timestamp.