Address the requested changes and add a new feature that the user can choose the number of comments displayed
Modify the index.html to add a new input box to let the user input the number of comments shown on the page.
Update the script.js file to show certain number of comments according to the users' input.
Modify the Comment.java and DataServlet.java files to address the requested changes.