Adjust the behavior of the input box for the maximum number of comments
Update the index.html file to let the onchange function of the input box and the onload function of the page to be different.
Modify the script.js file to add a new function for the onchange behavior of the input box. Update the DataServelt.java file
to store the maximum number of comments that the user input each time they submit the form.