Add a new feature that the user can delete all comments in datastore
Add a new DeleteDataServlet.java file to handle the post request. When it receives a post request, it will
delete all Comment entites in datastore and return an empty response.
Update the script.js file to add a new function deleteComments() that when the function is called,
it will send a POST request to /delete-data, which will be handled by DeleteDataServlet.java. After the post
request is sent, the comments on the website will be deleted.
Update the index.html file to add a button that when the button is clicked, the deleteComments() function will
be called.
Modify the style.css file to change the appearance of the newly-added button.