Add a new feature that the user can like a certain comment by clicking the button below it.
Update the script.js file to create a button each time I add a comment and add a new function
likeComments() to send a POST request each time a user like the comment.
Create the LikeDataServlet.java file to handle the POST request by adding the number of likes
of the entity with the given comment by one.
Modify the DataServlet.java file to sort all the comments first according to the number of likes.