Address most requested changes and make improvements to the design
Modify the LikeDataServlet.java file to add a filter to the query. Update the script.js file
to modify the behavior of button such that it will not automatically submit the form and reload. Also,
deleteComments function now send DELETE request, which is more appropriate. The name of the functions
in the file are also changed to be more representative and understandable.
Update the DataServlet.java file to change the JSON format that it outputs. Now I have one more field
that will return the stored maximum comments. Also I change the name of function getUserInput to
getMaxComments to be more specific.
Improve the DeketeDataServlet file by removing some unnecessary lines and adjust the URL naming based on
the REST style.