Update the script file to pass the key string of the comment's corresponding entity in the query string
Update the likeComments function such that the queryURL will now pass a HTML-safe key string with the POST request.
LikeDataServlet.java, DataServlet.java and Comment.java are modified to accomodate the change.
Modify the url name in the DeleteDataServlet.java file.
Add the type of the list element in the CommentResponse.java file.