Add a feature that the marker placed by the user will be stored in datastore
Update the script.js file to send a POST a request every time the user place a marker and store the latitude and longitude of the position in datastore.
Create Marker.java and MarkerServlet.java file to deal with the POST request and manage the entities of type "Marker" in datastore.