Add a new feature that the user can remove a marker on the map by clicking on the map
Update the script.js file to add a event listener to the marker such that when the marker is clicked, it will remove itself from the map and send a DELETE request to /markers.
Update the MarkerServlet.java file to handle the DELETE request. It will fitler according the latitude and longitude and then delete the requested entity. So now, when a marker is clicked, its corresponding entity in datastore will also be removed.