Add a search box to the map
Update the index.html file to include a search box at the top of the map.
Modify the initMap() function to generate a search box when the map is initiated. When search for a places, the result will be restricted such that the top
results will near the current viewport.