Add a new feature that the user will see different versions of the website depending on the login status
Create LoginStatusServlet.java file to handle the GET request asking for the current user login in status. If the user is logged in, return "true". Otherwise, return "false".
Create AuthenticationServlet.java file to set up the login and logout page for the users.
Update the index.html file to include the login and logout link.
Update the script.js file to hide the login link when the user has already logged in. When the user is not logged in, the logout link and comment form will be hided.