Add a new feature that the author of the comment will also be shown in the comments box
Modify the DataServlet.java file to retrieve the current user's email address when it handles the POST request and add the user's email address to the JSON string when it handles the GET request.
Update the Comment.java file to add a new field to store the email address of the author.
Update the script.js file such that comments will not be fetched if the user is not logged in.