Add support for users specifying display nicknames and the nicknames will be shown in front of the comments instead of email address
Modify the Comment.java file such that it now takes in the users' nickname instead of email address.
Update the AuthenticationServlet.java file to make the getNickname function public and solve the problem of redirecting websites. Now, after a user logs in, the user will be redirected to the nickname setup page.
Update the DataServlet.java file to pass the user's nickname with the user's comment together in the json string when it handles a GET request.
Create NicknameServlet.java file to handle the POST request sent by nickname.html and stored the users' nickname in datastore.
Modify the script.js file to show users' nickname on the screen instead of email address.