Add a new feature that the maxComment value is bound with the user such that each user can set their favorite maximum comments per page and the value will not change each time they login
Update the CommentResponse.java file to take in maxComment in type long instead of int.
Modify the AuthenticationServlet to adjust the welcome message shown on the screen.
Update the DataServlet.java file such that the maxComment value is stored in datastore and retrieved when the servlet handles a GET request.