Update the login page and fix the bug that the nickname of the comment author was not correct
Update the RetrieveNickname.java file to include one more method that retrieve the user's nickname based on email.
Update the AuthenticationServlet.java file to update the login page text.
Modify DataServlet.java file to retrieve the user's nickname based on his email instead of the userservice such that the function will not return the nickname of the current user.